Refactored the current code
App works exactly the same. This was an attempt to simplify the code for non-rubyists as much as possible by removing anything that might be deemed confusing or unnecessary.

All cucumber/rspec tests are still green.
