+++ app.rb
@@ -1,0 +1,7 @@
+require './models/storage'
+
+get '/' do
+  @holidays = Storage.find({})
+
+  erb :index
+end
