+++ features/holidays.feature
@@ -1,5 +1,5 @@
 Feature: See holidays listed
 
   Scenario: See holidays listed
-    When I go to the home page
+    Given I am on the home page
     Then I see the holidays listed
