+++ features/support/env.rb
@@ -1,21 +1,10 @@
-# Generated by cucumber-sinatra. (2015-12-26 21:17:42 +0000)
-
-ENV['RACK_ENV'] = 'test'
-
-require File.join(File.dirname(__FILE__), '..', '..', 'lib/holidays.rb')
-
+require 'sinatra'
 require 'capybara'
 require 'capybara/cucumber'
 require 'rspec'
+require './app'
 
-Capybara.app = Holidays::Application
+set :environment, :test
+set :views, './views'
 
-class Holidays::ApplicationWorld
-  include Capybara::DSL
-  include RSpec::Expectations
-  include RSpec::Matchers
-end
-
-World do
-  Holidays::ApplicationWorld.new
-end
+Capybara.app = Sinatra::Application
