+++ spec/app_spec.rb
@@ -1,0 +1,11 @@
+describe 'App' do
+  def app
+    Sinatra::Application
+  end
+
+  it 'responds successfully' do
+    get '/'
+
+    expect(last_response).to be_ok
+  end
+end
