+++ spec/models/storage_spec.rb
@@ -1,0 +1,13 @@
+describe Storage do
+  describe '#find' do
+    subject(:holidays) { Storage.find(criteria) }
+
+    context 'when given no criteria' do
+      let(:criteria) { {} }
+
+      it 'returns all holidays' do
+        expect(holidays.count).to eq(12)
+      end
+    end
+  end
+end
